% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_correlation_site.R
\name{auto_correlation_site}
\alias{auto_correlation_site}
\title{Compute an autocorrelation coefficient at a given site in the FLUXNET2015 dataset}
\usage{
auto_correlation_site(in_data, max_lag)
}
\arguments{
\item{in_data}{data frame of the site we are studying.  The first column contains the time variable}

\item{max_lag}{the maximum number of lags to be considered}
}
\value{
A data frame that reports the absolute value of the autocorrelation function and the variables associated with it
}
\description{
Given a site, determine if all possible combinations of the variables are autocorrelated to a maximum lag. This is a tidy version of the cross correlation matrix.
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from rEMD pacakge vignette
}
