% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm_site.R
\name{ccm_site}
\alias{ccm_site}
\title{Compute a convergent cross map test at a given site in the FLUXNET2015 dataset}
\usage{
ccm_site(in_data)
}
\arguments{
\item{in_data}{data frame of the site we are studying.  This is a nested data frame that contains}

\item{alpha}{the significance level}

\item{max_lag}{the maximum number of lags to be considered}
}
\value{
A data frame that reports the  value of the cross correlation mapping for a given measurement
}
\description{
Given a site, determine if all possible combinations of the variables are Granger cause each other. The Null Hypothesis (H0) is that Time series x does not Granger-cause time series y, and the Alternative Hypothesis (HA): Time series x Granger-causes time series y
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from Chandler Lutz, UCR 2009 (https://www.mathworks.com/matlabcentral/fileexchange/25467-granger-causality-test).
}
