% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix_construct.R
\name{correlation_matrix_construct}
\alias{correlation_matrix_construct}
\title{Construct a cross correlation matrix}
\usage{
correlation_matrix_construct(in_data, max_lag)
}
\arguments{
\item{max_lag}{the maximum number of lags to be considered}

\item{data}{incoming data frame}

\item{na_action}{what to do if we have NAs in data (default - omit)}
}
\value{
Cross correlation matrix
}
\description{
Given data frame compute the cross correlation between all variables, up to a given lag
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from https://cran.r-project.org/web/packages/rEDM/vignettes/rEDM-tutorial.pdf
}
