% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_dimension_site.R
\name{embedding_dimension_site}
\alias{embedding_dimension_site}
\title{Compute a the embedding dimension at a given site in the FLUXNET2015 dataset}
\usage{
embedding_dimension_site(in_data)
}
\arguments{
\item{in_data}{data frame of the site we are studying.  The first column contains the time variable}
}
\value{
A string (more here)
}
\description{
Given a site, determine the optimal embedding dimension so we can apply convergent cross mapping
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from Chandler Lutz, UCR 2009 (https://www.mathworks.com/matlabcentral/fileexchange/25467-granger-causality-test).
}
