% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger_site.R
\name{granger_site}
\alias{granger_site}
\title{Compute a Granger Causality test at a given site in the FLUXNET2015 dataset}
\usage{
granger_site(in_data, alpha, max_lag)
}
\arguments{
\item{in_data}{data frame of the site we are studying.  The first column contains the time variable}

\item{alpha}{the significance level}

\item{max_lag}{the maximum number of lags to be considered}
}
\value{
A data frame that reports the F statistic, the critical value, and if F > c_v.   (if F > c_v we reject the null hypothesis that x does not Granger Cause y)
}
\description{
Given a site, determine if all possible combinations of the variables are Granger cause each other. The Null Hypothesis (H0) is that Time series x does not Granger-cause time series y, and the Alternative Hypothesis (HA): Time series x Granger-causes time series y
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from Chandler Lutz, UCR 2009 (https://www.mathworks.com/matlabcentral/fileexchange/25467-granger-causality-test).
}
