% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger_test.R
\name{granger_test}
\alias{granger_test}
\title{Compute a Granger Causality test}
\usage{
granger_test(x, y, alpha, max_lag = 1)
}
\arguments{
\item{x}{Predictor variable (can be a vector)}

\item{y}{Response variable (can be a vector)}

\item{alpha}{the significance level}

\item{max_lag}{the maximum number of lags to be considered}
}
\value{
A data frame that reports the F statistic, the critical value, and if F > c_v.   (if F > c_v we reject the null hypothesis that x does not Granger Cause y)
}
\description{
Given an x (predictor) and y (response) variable, determine if x causes way at the given lag. Uses the F test statistic to compare/evaluate. The Null Hypothesis (H0) is that Time series x does not Granger-cause time series y, and the Alternative Hypothesis (HA): Time series x Granger-causes time series y
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from Chandler Lutz, UCR 2009 (https://www.mathworks.com/matlabcentral/fileexchange/25467-granger-causality-test).
}
