% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_thuille.R, R/la_thuille_small.R
\docType{data}
\name{la_thuille}
\alias{la_thuille}
\title{Measured fluxnet variables across the LaThuille 2015 dataset}
\format{
A data frame with 428083 rows and 10 variables

A data frame with 428083 rows and 10 variables. This is a smaller dataset than from the la_thuille data taking sites that have a full year's worth of data for analysis AND more than five years represented.  NOTE: this could mean that some measurements are not measured, but the data record is there.
}
\usage{
data(la_thuille)

data(la_thuille)
}
\description{
\itemize{
  \item years. Year of measurement
  \item day. Julian Day of the year
  \item precip. Precipitation (mm?)
  \item NEE. Net carbon uptake (gC m-2)
  \item GPP. Gross Primary Productivity (gC m-2)
  \item TER. Total Ecosystem Respiration
  \item airTemp. Air Temperature (degrees C)
  \item site. Fluxnet abbreviation
  \item X_flag. summary flag from half hour measurements 0 = bad, 1 = good
}

\itemize{
  \item years. Year of measurement
  \item day. Julian Day of the year
  \item precip. Precipitation (mm?)
  \item NEE. Net carbon uptake (gC m-2)
  \item GPP. Gross Primary Productivity (gC m-2)
  \item TER. Total Ecosystem Respiration
  \item airTemp. Air Temperature (degrees C)
  \item site. Fluxnet abbreviation
}
}
\keyword{datasets}
