% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_granger.R
\name{theme_granger}
\alias{theme_granger}
\title{Define a custom theme for making ggplots}
\usage{
theme_granger()
}
\description{
\code{theme_granger} Sets up defaults for ggplots so I don't need to keep doing these over and over.
}
