% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_entropy_site.R
\name{transfer_entropy_site}
\alias{transfer_entropy_site}
\title{Compute a transfer entropy test at a given site in the FLUXNET2015 dataset}
\usage{
transfer_entropy_site(in_data, alpha, max_lag, bootstrap_samples = 1)
}
\arguments{
\item{in_data}{data frame of the site we are studying.  The first column contains the time variable}

\item{alpha}{the significance level}

\item{max_lag}{the maximum number of lags to be considered}

\item{bootstrap_samples}{Number of bootstrap samples we apply}
}
\value{
A T/F value - this means that the conditional information is less than the significance level alpha for all the considered lags.
}
\description{
Given a site, determine if all possible combinations of the variables are Granger cause each other. The Null Hypothesis (H0) is that Time series x does not cause time series y, and the Alternative Hypothesis (HA): Time series x causes time series y
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from Chandler Lutz, UCR 2009 (https://www.mathworks.com/matlabcentral/fileexchange/25467-granger-causality-test).
}
