% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_entropy_test.R
\name{transfer_entropy_test}
\alias{transfer_entropy_test}
\title{Compute a Transfer Entropy Test}
\usage{
transfer_entropy_test(x, y, alpha = 0.05, max_lag = 1)
}
\arguments{
\item{x}{Predictor variable (can be a vector)}

\item{y}{Response variable (can be a vector)}

\item{alpha}{the significance level}

\item{max_lag}{the maximum number of lags to be considered}
}
\value{
A T/F value - this means that the conditional information is less than the significance level alpha for all the considered lags.
}
\description{
Given an x (predictor) and y (response) variable, determine if x causes way at the given lag. Uses conditional mutual information to assess The Null Hypothesis (H0) is that Time series x does not cause time series y, and the Alternative Hypothesis (HA): Time series x causes time series y
}
\author{
John Zobitz \email{zobitz@augsburg.edu}, based on code from Chandler Lutz, UCR 2009 (https://www.mathworks.com/matlabcentral/fileexchange/25467-granger-causality-test).
}
